fix: #420 signup debug
