+++ package.json
@@ -13,6 +13,8 @@
   "dependencies": {
     "@emotion/react": "^11.11.4",
     "@emotion/styled": "^11.11.5",
+    "@fortawesome/fontawesome-free": "^6.6.0",
+    "@mui/material": "^5.16.4",
     "async-validator": "^4.2.5",
     "axios": "^1.7.2",
     "bulma": "^1.0.1",
